MSDEV-2138 expose hooks via methods, move hooks to separate directory
