+++ xlogger/logger.go
@@ -116,8 +116,10 @@ func (l *Logger) Log(log *logrus.Logger) *logrus.Entry {
 		if l.cfg.HideFName {
 			return log.WithField("caller", caller)
 		}
+
 		return log.WithField("caller", caller).WithField("fName", fName)
 	}
+
 	return &logrus.Entry{}
 }
 
@@ -198,13 +200,12 @@ func (l *Logger) AddHook(hook logrus.Hook) {
 	l.log.Hooks.Add(hook)
 }
 
-func (l *Logger) AddHook2(hook logrus.Hook) {
-	l.log.Hooks.Add(hook)
-	l.errLog.Hooks.Add(hook)
+func (l *Logger) Hooks() logrus.LevelHooks {
+	return l.log.Hooks
 }
 
-func (l *Logger) AddErrHook(hook logrus.Hook) {
-	l.errLog.Hooks.Add(hook)
+func (l *Logger) ErrHooks() logrus.LevelHooks {
+	return l.errLog.Hooks
 }
 
 func (l *Logger) Dump(v ...interface{}) {
